+++ cmd/thanos/receive.go
@@ -1,20 +1,24 @@
 package main
 
 import (
-	"context"
 	"fmt"
+	stdlog "log"
 	"net"
+	"net/http"
 	"sync"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/improbable-eng/thanos/pkg/component"
+	"github.com/improbable-eng/thanos/pkg/prober"
 	"github.com/improbable-eng/thanos/pkg/receive"
 	"github.com/improbable-eng/thanos/pkg/runutil"
 	"github.com/improbable-eng/thanos/pkg/store"
 	"github.com/improbable-eng/thanos/pkg/store/storepb"
+	conntrack "github.com/mwitkow/go-conntrack"
 	"github.com/oklog/run"
+	"github.com/opentracing-contrib/go-stdlib/nethttp"
 	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
@@ -28,10 +32,7 @@ func registerReceive(m map[string]setupFunc, app *kingpin.Application, name stri
 	cmd := app.Command(name, "Accept Prometheus remote write API requests and write to local tsdb (EXPERIMENTAL, this may change drastically without notice)")
 
 	grpcBindAddr, cert, key, clientCA := regGRPCFlags(cmd)
-	httpMetricsBindAddr := regHTTPAddrFlag(cmd)
-
-	remoteWriteAddress := cmd.Flag("remote-write.address", "Address to listen on for remote write requests.").
-		Default("0.0.0.0:19291").String()
+	httpBindAddr := regHTTPAddrFlag(cmd)
 
 	dataDir := cmd.Flag("tsdb.path", "Data directory of TSDB.").
 		Default("./data").String()
@@ -46,8 +47,7 @@ func registerReceive(m map[string]setupFunc, app *kingpin.Application, name stri
 			*cert,
 			*key,
 			*clientCA,
-			*httpMetricsBindAddr,
-			*remoteWriteAddress,
+			*httpBindAddr,
 			*dataDir,
 		)
 	}
@@ -62,11 +62,10 @@ func runReceive(
 	cert string,
 	key string,
 	clientCA string,
-	httpMetricsBindAddr string,
-	remoteWriteAddress string,
+	httpBindAddr string,
 	dataDir string,
 ) error {
-	logger = log.With(logger, "component", "receive")
+	logger = log.With(logger, "component", component.Receive.String())
 	level.Warn(logger).Log("msg", "setting up receive; the Thanos receive component is EXPERIMENTAL, it may break significantly without notice")
 
 	tsdbCfg := &tsdb.Options{
@@ -78,16 +77,18 @@ func runReceive(
 
 	localStorage := &tsdb.ReadyStorage{}
 	receiver := receive.NewWriter(log.With(logger, "component", "receive-writer"), localStorage)
-	webHandler := receive.NewHandler(log.With(logger, "component", "receive-handler"), &receive.Options{
-		Receiver:      receiver,
-		ListenAddress: remoteWriteAddress,
-		Registry:      reg,
-		ReadyStorage:  localStorage,
-	})
 
 	// Start all components while we wait for TSDB to open but only load
 	// initial config and mark ourselves as ready after it completed.
 	dbOpen := make(chan struct{})
+
+	readinessProber := prober.NewProber(component.Receive.String(), logger)
+	webHandler := receive.NewHandler(log.With(logger, "component", "receive-handler"), &receive.Options{
+		Receiver:        receiver,
+		ReadinessProber: readinessProber,
+		Registry:        reg,
+		ReadyStorage:    localStorage,
+	})
 
 	// sync.Once is used to make sure we can close the channel at different execution stages(SIGTERM or when the config is loaded).
 	type closeOnce struct {
@@ -165,11 +166,6 @@ func runReceive(
 		)
 	}
 
-	level.Debug(logger).Log("msg", "setting up metric http listen-group")
-	if err := metricHTTPListenGroup(g, logger, reg, httpMetricsBindAddr); err != nil {
-		return err
-	}
-
 	level.Debug(logger).Log("msg", "setting up grpc server")
 	{
 		var (
@@ -210,20 +206,43 @@ func runReceive(
 		})
 	}
 
-	level.Debug(logger).Log("msg", "setting up receive http handler")
+	level.Debug(logger).Log("msg", "setting up http handler")
 	{
-		ctx, cancel := context.WithCancel(context.Background())
-		g.Add(
-			func() error {
-				if err := webHandler.Run(ctx); err != nil {
-					return fmt.Errorf("error starting web server: %s", err)
-				}
-				return nil
-			},
-			func(err error) {
-				cancel()
-			},
-		)
+
+		mux := http.NewServeMux()
+		registerMetrics(mux, reg)
+		registerProfile(mux)
+		readinessProber.RegisterInMux(mux)
+		webHandler.HandleInMux("/", mux)
+
+		listener, err := net.Listen("tcp", httpBindAddr)
+		if err != nil {
+			return errors.Wrapf(err, "listen HTTP on address %s", httpBindAddr)
+		}
+
+		// Monitor incoming connections with conntrack.
+		listener = conntrack.NewListener(listener,
+			conntrack.TrackWithName("http"),
+			conntrack.TrackWithTracing())
+
+		operationName := nethttp.OperationNameFunc(func(r *http.Request) string {
+			return fmt.Sprintf("%s %s", r.Method, r.URL.Path)
+		})
+
+		errlog := stdlog.New(log.NewStdlibAdapter(level.Error(logger)), "", 0)
+
+		httpSrv := &http.Server{
+			Handler:  nethttp.Middleware(opentracing.GlobalTracer(), mux, operationName),
+			ErrorLog: errlog,
+		}
+
+		g.Add(func() error {
+			level.Info(logger).Log("msg", "Listening for query and metrics", "address", httpBindAddr)
+			webHandler.Healthy()
+			return httpSrv.Serve(listener)
+		}, func(err error) {
+			runutil.CloseWithLogOnErr(logger, listener, "receive HTTP listener")
+		})
 	}
 	level.Info(logger).Log("msg", "starting receiver")